% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_bio_by_species.R
\name{cleanup_bio_by_species}
\alias{cleanup_bio_by_species}
\title{Cleanup bio data species function}
\usage{
cleanup_bio_by_species(
  bio_df,
  catch_cleaned,
  filter_lat_lt,
  filter_lat_gt,
  filter_depth,
  species
)
}
\arguments{
\item{bio_df}{Dataframe of biological data - length and age compositions that
has a Trawl_id column}

\item{catch_cleaned}{output of cleanup_by_species}

\item{filter_lat_lt}{numeric value for latitude to filter out of the biological data}

\item{filter_lat_gt}{numeric value for latitude to filter out of the biological data}

\item{filter_depth}{numeric value for depth to filter out of the biological data}

\item{species}{character vector of the species name to filter out of all the 
biological data if more than one species}
}
\value{
List of resampled biological dataframes that match the catch data
}
\description{
Filter biological by species
}
