% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_by_species.R
\name{join_dfs}
\alias{join_dfs}
\title{Join Data Frames}
\usage{
join_dfs(list_of_dfs, main_df, shared_column)
}
\arguments{
\item{list_of_dfs}{The list of data frames that you would like to join to
the main data frame}

\item{main_df}{The main data frame that you want to join the list of data
frames}

\item{shared_column}{The column that all the data frames share which will be
used to join the data frames together.}
}
\value{
A list of merged dataframes by the shared_column
}
\description{
Function to join a list of data frames to a main data frame using a shared
column.
}
