% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_nefsc_pred_grid_depth.R
\docType{data}
\name{noaa_nefsc_pred_grid_depth}
\alias{noaa_nefsc_pred_grid_depth}
\title{Prediction grid for NWA survey}
\format{
A data frame with 80413 observations on the following 6 variables.
\describe{
\item{\code{srvy}}{Abbreviated survey names. Abbreviated survey names.}
\item{\code{latitude_dd}}{Latitude (decimal degrees). Latitude (one
hundred thousandth of a decimal degree).}
\item{\code{longitude_dd}}{Longitude (decimal degrees). Longitude (one
hundred thousandth of a decimal degree).}
\item{\code{pass}}{Pass. Pass}
\item{\code{depth_m}}{Depth (m). Bottom depth (meters).}
\item{\code{area_km2}}{Area (km2). Area in square kilometers.}
}
}
\source{
Data request to the NEFSC.
}
\usage{
data('noaa_nefsc_pred_grid_depth')
}
\description{
ENTER.
}
\details{
ENTER.
}
\examples{
data(noaa_nefsc_pred_grid_depth)
}
\author{
Northeast Fisheries Science Center, compiled by Emily Markowitz
(Emily.Markowitz AT noaa.gov)
}
\keyword{code}
\keyword{data}
\keyword{species}
