% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_by_species.R
\name{tow_fn}
\alias{tow_fn}
\title{Tow Function}
\usage{
tow_fn(x)
}
\arguments{
\item{x}{catch_split data frame}
}
\value{
data frame of tows to include or exclude
}
\description{
Create a vector of tows for including or excluding.
}
