% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_sdmtmb.R
\name{wrapper_sdmtmb}
\alias{wrapper_sdmtmb}
\title{Species distribution model function using sdmTMB}
\usage{
wrapper_sdmtmb(x, y, z, dir_spp, spp_info, n_knots = 300)
}
\arguments{
\item{x}{speciesname_df[[i]] which is a data frame from a list of data frames
created from the cleanup_by_species() function and any further
post-processing of depth filters (see the smaller_function.R file for those).}

\item{y}{speciesname_files[[i]] which is an item in a list created from}

\item{z}{A dataframe with the new data to predict on.}

\item{dir_spp}{A character string specifying the directory for output files.}

\item{spp_info}{information about the species test run.}

\item{n_knots}{Numeric.
names(speciesname_df)}
}
\description{
Function to create a mesh, fit the sdmTMB model, and get the index.
Exports fit.rds and index.rds files to the designated species folder.
Learn more: https://github.com/pbs-assess/sdmTMB/
}
\examples{
\dontrun{
wrapper_sdmtmb()
}
}
