% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_resample.R
\name{clean_and_resample}
\alias{clean_and_resample}
\title{Clean and Resample Species Data}
\usage{
clean_and_resample(
  spp_info,
  catch,
  bio,
  seq_from,
  seq_to,
  seq_by,
  tot_dataframes,
  replicate_num,
  grid_yrs,
  dir_out,
  test = FALSE,
  n_knots = 300,
  model_type = "wrapper_sdmtmb"
)
}
\arguments{
\item{spp_info}{A data frame row containing information about the species.}

\item{catch}{A data frame containing the catch data.}

\item{bio}{A data frame containing the biological data, if applicable.}

\item{seq_from}{A numeric value specifying the start of the sequence for data frames.}

\item{seq_to}{A numeric value specifying the end of the sequence for data frames.}

\item{seq_by}{A numeric value specifying the step size of the sequence for data frames.}

\item{tot_dataframes}{An integer specifying the total number of data frames to generate.}

\item{replicate_num}{An integer specifying the number of replicates.}

\item{grid_yrs}{A data frame or list containing grid years information.}

\item{dir_out}{A character string specifying the directory for output files.}

\item{test}{Logical. TRUE/FALSE. If TRUE, will only run first two resampling tests.}

\item{n_knots}{Numeric. Default  = 500.}

\item{model_type}{String. Default = "wrapper_sdmtmb", but can be any preset wrapper_*() function or a premade home built function.}
}
\value{
A list of data frames containing the cleaned and resampled catch data.
}
\description{
This function cleans up the catch data for a specific species and then performs 
resampling tests,
}
\details{
This function performs the following steps:
\itemize{
  \item Cleans up the catch data for the specified species using `cleanup_by_species`.
  \item Performs resampling tests on the cleaned data using `resample_tests`.
}
}
\examples{
dir_out <- here::here("vignettes", "output")
catch <- surveyresamplr::noaa_nwfsc_catch
grid_yrs <- replicate_df(dat = surveyresamplr::noaa_nwfsc_catch, time_name = "year", 
                         time_values = unique(catch$year))
spp_list <- data.frame(srvy = "CA",
                       common_name = "arrowtooth flounder", 
                       file_name = "arrowtooth_flounder", 
                       filter_lat_gt = 34,
                       filter_lat_lt = NA,
                       filter_depth = NA,
                       model_fn = "total_catch_wt_kg ~ 0 + factor(year) + pass",
                       model_family = "delta_gamma",
                       model_anisotropy = TRUE
                       model_spatiotemporal = c("iid, iid")
                      )
clean_and_resample(spp_info = spp_list, 
                   catch, 
                   seq_from = 0.1, 
                   seq_to = 1, 
                   seq_by = 0.1, 
                   tot_dataframes = 91, 
                   replicate_num = 10, 
                   grid_yrs = grid_yrs, 
                   dir_out = dir_out))

}
