% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_by_species.R
\name{cleanup_by_species}
\alias{cleanup_by_species}
\title{Cleanup species function}
\usage{
cleanup_by_species(
  catch,
  spp_info,
  seq_from = 0.1,
  seq_to = 1,
  seq_by = 0.1,
  tot_dataframes = 91,
  replicate_num = 10
)
}
\arguments{
\item{catch}{full catch df}

\item{spp_info}{A data frame containing information about the test species.}

\item{seq_from}{A numeric value specifying the start of the sequence for data frames.}

\item{seq_to}{A numeric value specifying the end of the sequence for data frames.}

\item{seq_by}{A numeric value specifying the step size of the sequence for data frames.}

\item{tot_dataframes}{the number of data frames you want to output. effort x 
replicates - (replicates - 1). 5x3-2}

\item{replicate_num}{An integer specifying the number of replicates.}
}
\value{
List of resampled catch dataframes
}
\description{
Filter catch by species, create a vector of tows and give tows a random
assignment to be resampled,
}
