% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_or_exclude.R
\name{include_or_exclude}
\alias{include_or_exclude}
\title{Include or Exclude}
\usage{
include_or_exclude(df, proportions, replicate_num)
}
\arguments{
\item{df}{tows data frame}

\item{proportions}{proportions developed using: props <- as.data.frame(seq
(0.1,1.0, by = 0.1)) replicated by the length of the tows dataframe. The name
of the props is "trawlid".}

\item{replicate_num}{going to be 10 for NWFSC and 3 for AK}
}
\description{
Specify how to downsample. For simple random sampling, a proportion of
stations should do.
}
