% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_or_exclude.R
\name{include_or_exclude}
\alias{include_or_exclude}
\title{Include or Exclude}
\usage{
include_or_exclude(df, proportions, replicate_num)
}
\arguments{
\item{df}{tows data frame}

\item{proportions}{proportions developed using: props <- as.data.frame(seq
(0.1,1.0, by = 0.1)) replicated by the length of the tows dataframe. The name
of the props is "trawlid".}

\item{replicate_num}{going to be 10 for NWFSC and 3 for AK}
}
\value{
List of dataframes with random assignments of which data to include
}
\description{
Specify how to downsample using output of tow_fn. For simple random sampling, 
a proportion of tows should do. Utilized in the cleanup_by_species function.
}
