% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_nwfsc_catch.R
\docType{data}
\name{noaa_nwfsc_catch}
\alias{noaa_nwfsc_catch}
\title{Prediction grid for NWFSC survey}
\format{
A data frame with 200070 observations on the following 9 variables.
\describe{
  \item{\code{}}{. }#'   }
}
\source{
Northwest Fisheries Science Center; https://raw.githubusercontent.com/pfmc-assessments/indexwc/main/data/california_current_grid.rda.
}
\usage{
data('noaa_nwfsc_catch')
}
\description{
[ENTER].
}
\details{
[ENTER].
}
\examples{
data(noaa_nwfsc_catch)
}
\author{
Northwest Fisheries Science Center, compiled by Emily Markowitz (Emily.Markowitz AT noaa.gov) and Derek Bolser (Derek.Bolser AT noaa.gov)
}
\keyword{code}
\keyword{data}
\keyword{species}
