% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_nwfsc_pred_grid_depth.R
\docType{data}
\name{noaa_nwfsc_pred_grid_depth}
\alias{noaa_nwfsc_pred_grid_depth}
\title{Prediction grid for NWFSC survey}
\format{
A data frame with 45628 observations on the following 6 variables.
\describe{
  \item{\code{srvy}}{Abbreviated survey names. Abbreviated survey names. }
  \item{\code{latitude_dd}}{Latitude (decimal degrees). Latitude (one hundred thousandth of a decimal degree).}
  \item{\code{longitude_dd}}{Longitude (decimal degrees). Longitude (one hundred thousandth of a decimal degree).}
  \item{\code{pass}}{Pass. Pass}
  \item{\code{depth_m}}{Depth (m). Bottom depth (meters).}
  \item{\code{area_km2}}{Area (km2). Area in square kilometers.}#'   }
}
\source{
Northwest Fisheries Science Center; https://raw.githubusercontent.com/pfmc-assessments/indexwc/main/data/california_current_grid.rda.
}
\usage{
data('noaa_nwfsc_pred_grid_depth')
}
\description{
ENTER.
}
\details{
ENTER.
}
\examples{
data(noaa_nwfsc_pred_grid_depth)
}
\author{
Northwest Fisheries Science Center, compiled by Emily Markowitz (Emily.Markowitz AT noaa.gov) and Derek Bolser (Derek.Bolser AT noaa.gov)
}
\keyword{code}
\keyword{data}
\keyword{species}
