% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot Results and Save Figures}
\usage{
plot_results(srvy, dir_out, dir_final = NULL)
}
\arguments{
\item{srvy}{A character string specifying the survey identifier.}

\item{dir_out}{A character string specifying the directory for output files.}
}
\description{
This function compiles results from species distribution models (SDMs), generates plots, and saves the figures.
}
\details{
This function performs the following steps:
\itemize{
  \item Creates a directory for saving images.
  \item Searches for relevant files in the output directory.
  \item Compiles data from the found files.
  \item Saves the compiled data into CSV files.
  \item Generates various plots such as boxplots and time series of biomass estimates.
  \item Saves the generated plots as PNG files.
  \item Saves the list of plots in an RData file.
}
}
