% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot Results and Save Figures}
\usage{
plot_results(srvy, dir_out, dir_final = NULL)
}
\arguments{
\item{srvy}{A character string specifying the survey identifier.}

\item{dir_out}{A character string specifying the directory for output files.}

\item{dir_final}{A character string specifying the directory for new result files.}
}
\value{
A list of plots and tables generated during the function execution.
}
\description{
This function compiles results from species distribution models (SDMs) output 
of the clean_and_resample function for which it generates plots, and saves the 
figures.
}
\details{
This function performs the following steps:
\itemize{
  \item Creates a directory for saving images.
  \item Searches for relevant files in the output directory.
  \item Compiles data from the found files.
  \item Saves the compiled data into CSV files.
  \item Generates various plots such as boxplots and time series of biomass estimates.
  \item TO DO: CREATE BIO PLOTS
  \item Saves the generated plots as PNG files.
  \item Saves the list of plots in an RData file.
}
}
