% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_tests.R
\name{resample_tests}
\alias{resample_tests}
\title{Resample Tests and Run SDM Processing}
\usage{
resample_tests(spp_dfs, spp_info, grid_yrs, dir_out, test = FALSE)
}
\arguments{
\item{spp_dfs}{A list of species data frames.}

\item{spp_info}{A data frame containing information about the test species.}

\item{grid_yrs}{A data frame or list containing grid years information.}

\item{dir_out}{A character string specifying the directory for output files.}

\item{test}{Logical. TRUE/FALSE. If TRUE, will only run first two resampling tests. 

#' @details
This function performs the following steps:
\itemize{
  \item Sets up directories for output files.
  \item Reduces the list of data frames to the last two entries for testing purposes.
  \item Saves each data frame in Parquet format.
  \item Sets up parallel processing using the \code{furrr} package.
  \item Runs species distribution models (SDMs) in parallel.
  \item Saves the results of the SDM processing into CSV files.
}}
}
\description{
This function resamples species data frames, runs species distribution models (SDMs) in parallel, and saves the results.
}
